requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.1.1
django-mysql==4.7.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.0
mysqlclient==2.1.1
PyJWT==1.7.1
pytz==2022.2.1
sqlparse==0.4.2